Fix(api): ajuste para recebimento via API

+++ backend/src/jobs/SendMessageAPI.ts
@@ -10,6 +10,10 @@ import { getWbot } from "../libs/wbot";
 import UpsertMessageAPIService from "../services/ApiMessageService/UpsertMessageAPIService";
 import Queue from "../libs/Queue";
 import CheckIsValidContact from "../services/WbotServices/CheckIsValidContact";
+import AppError from "../errors/AppError";
+import VerifyContact from "../services/WbotServices/helpers/VerifyContact";
+import FindOrCreateTicketService from "../services/TicketServices/FindOrCreateTicketService";
+import CreateMessageSystemService from "../services/MessageServices/CreateMessageSystemService";
 
 export default {
   key: "SendMessageAPI",
@@ -25,13 +29,10 @@ export default {
   async handle({ data }: any) {
     try {
       const wbot = getWbot(data.sessionId);
-      let message: any = {} as WbotMessage;
-
+      const message: any = {} as WbotMessage;
       try {
-        const waNumber = await CheckIsValidContact(data.number, data.tenantId);
-        data.number = waNumber.user;
-      } catch (error) {
-        if (error.message === "ERR_WAPP_INVALID_CONTACT") {
+        const idNumber = await wbot.getNumberId(data.number);
+        if (!idNumber) {
           const payload = {
             ack: -1,
             body: data.body,
@@ -39,11 +40,15 @@ export default {
             number: data.number,
             externalKey: data.externalKey,
             error: "number invalid in whatsapp",
-            type: "hookMessageStatus"
+            type: "hookMessageStatus",
+            authToken: data.authToken
           };
 
-          // excluir o arquivo se o número não existir
-          fs.unlinkSync(data.media.path);
+          if (data.media) {
+            // excluir o arquivo se o número não existir
+            fs.unlinkSync(data.media.path);
+          }
+
           if (data?.apiConfig?.urlMessageStatus) {
             Queue.add("WebHooksAPI", {
               url: data.apiConfig.urlMessageStatus,
@@ -53,103 +58,70 @@ export default {
           }
           return payload;
         }
+
+        // '559891191708@c.us'
+        const msgContact = await wbot.getContactById(idNumber._serialized);
+        const contact = await VerifyContact(msgContact, data.tenantId);
+        const ticket = await FindOrCreateTicketService({
+          contact,
+          whatsappId: wbot.id!,
+          unreadMessages: 0,
+          tenantId: data.tenantId,
+          groupContact: undefined,
+          msg: data,
+          channel: "whatsapp"
+        });
+
+        await CreateMessageSystemService({
+          msg: data,
+          tenantId: data.tenantId,
+          ticket,
+          sendType: "API",
+          status: "pending"
+        });
+
+        await ticket.update({
+          apiConfig: {
+            ...data.apiConfig,
+            externalKey: data.externalKey
+          }
+        });
+      } catch (error) {
+        const payload = {
+          ack: -2,
+          body: data.body,
+          messageId: "",
+          number: data.number,
+          externalKey: data.externalKey,
+          error: "error session",
+          type: "hookMessageStatus",
+          authToken: data.authToken
+        };
+
+        if (data?.apiConfig?.urlMessageStatus) {
+          Queue.add("WebHooksAPI", {
+            url: data.apiConfig.urlMessageStatus,
+            type: payload.type,
+            payload
+          });
+        }
         throw new Error(error);
       }
 
-      if (data.mediaUrl) {
-        try {
-          const request = await axios.get(data.mediaUrl, {
-            responseType: "stream"
-          });
-          const cType = request.headers["content-type"];
-          const fileExt = mime.extension(cType);
-          const mediaName = uuid();
-          const path =
-            process.env.PATH_OFFLINE_MEDIA ||
-            join(__dirname, "..", "..", "..", "public");
-          const mediaDir = `${path}/${mediaName}`;
-          const fileName = `${mediaName}.${fileExt}`;
-          fs.mkdirSync(mediaDir);
-          const mediaPath = join(mediaDir, fileName);
-          await new Promise((resolve, reject) => {
-            request.data
-              .pipe(fs.createWriteStream(mediaPath))
-              .on("finish", async () => {
-                const newMedia = MessageMedia.fromFilePath(mediaPath);
-                message = await wbot.sendMessage(
-                  `${data.number}@c.us`,
-                  newMedia,
-                  {
-                    sendAudioAsVoice: true,
-                    caption: data.body
-                  }
-                );
-                fs.unlinkSync(mediaPath);
-                fs.rmdirSync(mediaDir);
-                resolve(message);
-              })
-              .on("error", (error: any) => {
-                console.error("ERROR DONWLOAD", error);
-                fs.rmdirSync(mediaDir, { recursive: true });
-                reject(new Error(error));
-              });
-          });
-        } catch (error) {
-          if (error.response.status === 404) {
-            const payload = {
-              ack: -1,
-              body: data.body,
-              messageId: "",
-              number: data.number,
-              externalKey: data.externalKey,
-              error: error.message,
-              type: "hookMessageStatus"
-            };
-            if (data?.apiConfig?.urlMessageStatus) {
-              Queue.add("WebHooksAPI", {
-                url: data.apiConfig.urlMessageStatus,
-                type: payload.type,
-                payload
-              });
-            }
-            return payload;
-          }
-          throw new Error(error);
-        }
-      }
-
-      if (data.media && !data.mediaUrl) {
-        const newMedia = MessageMedia.fromFilePath(data.media.path);
-        message = await wbot.sendMessage(`${data.number}@c.us`, newMedia, {
-          sendAudioAsVoice: true,
-          caption: data.body
-        });
-
-        fs.unlinkSync(data.media.path);
-      }
-
-      if (!data.media && !data.mediaUrl) {
-        message = await wbot.sendMessage(`${data.number}@c.us`, data.body, {
-          linkPreview: false
-        });
-      }
-
-      const apiMessage = await UpsertMessageAPIService({
-        sessionId: data.sessionId,
-        messageId: message.id.id,
-        body: data.body,
-        ack: message.ack,
-        number: data.number,
-        mediaName: data?.media?.filename,
-        mediaUrl: data.mediaUrl,
-        timestamp: message.timestamp,
-        externalKey: data.externalKey,
-        messageWA: message,
-        apiConfig: data.apiConfig,
-        tenantId: data.tenantId
-      });
-
-      return apiMessage;
+      // const apiMessage = await UpsertMessageAPIService({
+      //   sessionId: data.sessionId,
+      //   messageId: message.id.id,
+      //   body: data.body,
+      //   ack: message.ack,
+      //   number: data.number,
+      //   mediaName: data?.media?.filename,
+      //   mediaUrl: data.mediaUrl,
+      //   timestamp: message.timestamp,
+      //   externalKey: data.externalKey,
+      //   messageWA: message,
+      //   apiConfig: data.apiConfig,
+      //   tenantId: data.tenantId
+      // });
     } catch (error) {
       logger.error({ message: "Error send message api", error });
       throw new Error(error);
